- support HMAC flag which if present in the attribute list toggles to HMAC authentication transforms and otherwise to simple keyed authentication. Note, HMAC is necessary if new esp is to use integrity checking, i.e. authentication of the payload. - Also fix bug, where SPIs were reserved for more than one protocol when only one protocol, e.g. ESP or AH, could be agreed upon. - Also make kernel.c a bit less complex, I hope. - return notifies to kernel on failure only when kernel started the keying.

+++ src/sbin/ipsec/photurisd/attributes.c
@@ -1,5 +1,5 @@
 /*
- * Copyright 1997 Niels Provos <provos@physnet.uni-hamburg.de>
+ * Copyright 1997,1998 Niels Provos <provos@physnet.uni-hamburg.de>
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -61,9 +61,9 @@ putattrib(attrib_t *attrib)
 }
 
 attrib_t *
-getattrib(int id) 
+getattrib(u_int8_t id) 
 {
-     int hashval = id % ATTRIBHASHMOD;
+     u_int8_t hashval = id % ATTRIBHASHMOD;
      attrib_t *attrib;
 
      for(attrib=attribhash[hashval]; attrib; attrib = attrib->next)
